renamed `ModDropLocations to DropLocations`
